Move telemetry code into own module
This allows further enhancements to be encapsulated from the server
implementation.
