+++ go.mod
@@ -7,7 +7,7 @@ require (
 	github.com/containers/common v0.50.1
 	github.com/containers/storage v1.43.0
 	github.com/google/uuid v1.3.0
-	github.com/onsi/ginkgo/v2 v2.2.0
+	github.com/onsi/ginkgo/v2 v2.3.0
 	github.com/onsi/gomega v1.21.1
 	github.com/opencontainers/runc v1.1.4
 	github.com/opencontainers/runtime-tools v0.9.1-0.20220714195903-17b3287fafb7
